@@ -2,22 +2,22 @@
   "name": "relay-reactjs-chat",
   "version": "0.1.0",
   "private": true,
-  "homepage": "/examples/relay-reactjs-chat",
+  "homepage": "/relay-reactjs-chat",
   "dependencies": {
-    "@testing-library/jest-dom": "^5.16.2",
-    "@testing-library/react": "^13.2.0",
-    "@testing-library/user-event": "^14.0.4",
-    "js-waku": "^0.24.0",
-    "protobufjs": "^6.11.2",
-    "react": "^18.1.0",
-    "react-dom": "^18.1.0",
+    "@testing-library/jest-dom": "^5.16.5",
+    "@testing-library/react": "^13.3.0",
+    "@testing-library/user-event": "^13.5.0",
+    "js-waku": "0.24.0-63bfb9b",
+    "protobufjs": "^7.0.0",
+    "react": "^18.2.0",
+    "react-dom": "^18.2.0",
     "react-scripts": "5.0.1"
   },
   "scripts": {
-    "start": "cra-webpack-rewired start",
-    "build": "cra-webpack-rewired build",
-    "test": "cra-webpack-rewired test",
-    "eject": "cra-webpack-rewired eject"
+    "start": "craco start",
+    "build": "craco build",
+    "test": "craco test",
+    "eject": "craco eject"
   },
   "eslintConfig": {
     "extends": [
@@ -40,9 +40,6 @@
     ]
   },
   "devDependencies": {
-    "cra-webpack-rewired": "^1.0.1",
-    "process": "^0.11.10",
-    "stream-browserify": "^3.0.0",
-    "typescript": "^4.7.4"
+    "@craco/craco": "7.0.0-alpha.7"
   }
 }
